Update release_tests_prod.py
Add param for release directory.

+++ tests/release_tests/release_tests_prod.py
@@ -1,3 +1,4 @@
+import os
 import re
 import csv
 import json
@@ -11,6 +12,7 @@ from time import sleep
 from bs4 import BeautifulSoup
 from selenium import webdriver
 from selenium.webdriver.firefox.options import Options
+
 
 options = Options()
 options.headless = True
@@ -99,7 +101,11 @@ def compare_random(compare_ids):
     return diff_response_ids
 
 
-def check_release_files(all_ror_ids_file, release_tests_outfile, jsondiff_outfile):
+def check_release_files(release_directory, all_ror_ids_file, release_tests_outfile, jsondiff_outfile):
+    if not os.path.exists(release_directory):
+        print(f"Error: Release directory '{release_directory}' does not exist.")
+        exit(1)
+
     with open(release_tests_outfile, 'w') as f_out:
         writer = csv.writer(f_out)
         writer.writerow(["ror_id", "org_name", "retrieve_check",
@@ -108,7 +114,8 @@ def check_release_files(all_ror_ids_file, release_tests_outfile, jsondiff_outfil
         writer = csv.writer(f_out)
         writer.writerow(["ror_id", "diff"])
     ids_in_file = []
-    for file in glob.glob("*.json"):
+    json_files = glob.glob(os.path.join(release_directory, "*.json"))
+    for file in json_files:
         with open(file, 'r+', encoding='utf8') as f_in:
             json_file = json.load(f_in)
         ids_in_file.append(json_file["id"])
@@ -141,6 +148,8 @@ def check_release_files(all_ror_ids_file, release_tests_outfile, jsondiff_outfil
 
 def parse_arguments():
     parser = argparse.ArgumentParser(description='Run release tests for ROR')
+    parser.add_argument('-r', '--release_directory', required=True,
+                        help='Path to the directory containing JSON files')
     parser.add_argument('-a', '--all_ror_ids_file', default='all_ror_ids.txt',
                         help='Path to the release tests output file')
     parser.add_argument('-t', '--release_tests_outfile', default='release_tests.csv',
@@ -153,7 +162,7 @@ def parse_arguments():
 
 def main():
     args = parse_arguments()
-    check_release_files(args.all_ror_ids_file,
+    check_release_files(args.release_directory, args.all_ror_ids_file,
                         args.release_tests_outfile, args.jsondiff_outfile)
